handset tlm: values converted to correct format

+++ src/src/rc_channels.h
@@ -197,7 +197,7 @@ RcChannels_link_stas_pack(uint8_t *const output,
 void ICACHE_RAM_ATTR
 RcChannels_link_stas_extract(uint8_t const *const input,
                              LinkStats_t &output,
-                             int8_t snr, uint8_t rssi);
+                             int8_t snr, int16_t rssi);
 
 
 /*************************************************************************************
